+++ app/src/main/java/com/jpp/mp/main/movies/MovieListItem.kt
@@ -1,0 +1,13 @@
+package com.jpp.mp.main.movies
+
+/**
+ * Represents an item in the list of Movies.
+ */
+data class MovieListItem(
+    val movieId: Double,
+    val headerImageUrl: String,
+    val title: String,
+    val contentImageUrl: String,
+    val popularity: String,
+    val voteCount: String
+)
